+++ demo/meeting/src/containers/MeetingRoster.tsx
@@ -1,14 +1,14 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React from 'react';
+import React, { useState, ChangeEvent } from 'react';
 import styled from 'styled-components';
 import {
   Roster,
   RosterHeader,
-  RosterGroup
+  RosterGroup,
+  useRoster
 } from 'amazon-chime-sdk-component-library-react';
-import { useRoster } from '../../../../src';
 
 import RosterAttendee from './RosterAttendee';
 
@@ -22,8 +22,21 @@ const StyledRoster = styled.div`
 
 const MeetingRoster = () => {
   const roster = useRoster();
+  const [filter, setFilter] = useState('');
 
-  const attendees = Object.values(roster).map(item => {
+  let attendees = Object.values(roster);
+
+  if (filter) {
+    attendees = attendees.filter(attendee =>
+      attendee.name.toLowerCase().includes(filter.toLowerCase())
+    );
+  }
+
+  const handleSearch = (e: ChangeEvent<HTMLInputElement>) => {
+    setFilter(e.target.value);
+  };
+
+  const attendeeItems = attendees.map(item => {
     const { id, name } = item;
     return <RosterAttendee key={id} attendeeId={id} name={name} />;
   });
@@ -31,8 +44,14 @@ const MeetingRoster = () => {
   return (
     <StyledRoster>
       <Roster>
-        <RosterHeader title="Present" badge={attendees.length} />
-        <RosterGroup>{attendees}</RosterGroup>
+        <RosterHeader
+          searchValue={filter}
+          onSearch={handleSearch}
+          onClose={() => alert('Closing')}
+          title="Present"
+          badge={attendees.length}
+        />
+        <RosterGroup>{attendeeItems}</RosterGroup>
       </Roster>
     </StyledRoster>
   );