+++ demo/meeting/webpack.config.js
@@ -29,7 +29,11 @@ module.exports = {
     alias: {
       react: path.resolve('./node_modules/react'),
       'styled-components': path.resolve('./node_modules/styled-components'),
-      'react-dom': path.resolve('./node_modules/react-dom')
+      'react-dom': path.resolve('./node_modules/react-dom'),
+      'amazon-chime-sdk-component-library-react': path.resolve(
+        __dirname,
+        '../../src'
+      )
     }
   },
   output: {